v1.0.0: full SDK rewrite
Complete redesign of the Oddin Go SDK: a modern, bug-fixed replacement
for the v0.0.x codebase. Public surface is the ctx-aware Client with
Subscribe/Connect/Replay, typed SessionMessage envelopes, lossy
observability event channels, and errors.Is-able sentinels
(ErrInvalidConfig, ErrAlreadyClosed) plus a typed api.Error.

Highlights over v0.0.x:
- AMQP consumed with manual ack (noAck=false): deliveries are acked only
  after the decoded message is admitted to the public subscription
  buffer, the ack riding the pipeline (consumer -> session -> pump);
  broker prefetch is the real backpressure knob. Graceful vs abrupt
  close honour the drain contract; nothing is acked-then-lost.
  snapshot_complete is a correctness event: admitted to the recovery
  actor with backpressure and acked only after admission (else left for
  redelivery), so a full inbox can't strand recovery.
- Bounded, ctx-aware caches with singleflight de-duplication, per-key +
  purge-generation clear tombstones, atomic side-state (icon) commit on
  parent admission, and construction/Close lifetime cancellation of
  detached loads.
- Recovery state machine with reliable per-request completion handles,
  correct message-gen cursor advancement, initial-snapshot lookback, and
  request IDs bounded to positive int32 range (32-bit safe).
- backoff/v5 with explicit unlimited MaxElapsedTime on reconnect/API
  retry; Go 1.26 toolchain for stdlib CVE patches.
- CompositeKey fixed from *string (pointer-identity) to value key,
  closing the unbounded market-description cache growth.
- Producer catalog refreshes reuse the same *data so handles stay live;
  availability read under the data lock; ProducerScope is a wire-matching
  typed string with a ProducerHasScope free-function helper.
- SDK version reporting: every API request carries User-Agent +
  X-Oddin-SDK-Version and each AMQP connection reports SDK/SDK_version;
  the version resolves at runtime from build info with a -dev fallback,
  and a CI guard (golang.org/x/mod validation) keeps the release tag and
  version constant in lockstep.
- New validates config up front (bounded options + structural host form)
  as ErrInvalidConfig, instead of failing at a later API/AMQP call.

Release identity: module path stays github.com/oddin-gg/gosdk (no /v2),
so the first stable tag is v1.0.0. NEXT.md is the design source of
truth; MIGRATION.md documents the breaking changes for consumers.

+++ internal/feed/xml/msg_rollback_bet_cancel.go
@@ -9,8 +9,8 @@ import (
 type RollbackBetCancel struct {
 	MessageAttributes
 	XMLName   xml.Name                  `xml:"rollback_bet_cancel"`
-	StartTime *uint                     `xml:"start_time,attr,omitempty"`
-	EndTime   *uint                     `xml:"end_time,attr,omitempty"`
+	StartTime *int64                    `xml:"start_time,attr,omitempty"`
+	EndTime   *int64                    `xml:"end_time,attr,omitempty"`
 	Markets   []RollbackBetCancelMarket `xml:"market"`
 }
 
@@ -25,7 +25,7 @@ func (o RollbackBetCancel) GetEventID() string {
 }
 
 // Product ...
-func (o RollbackBetCancel) Product() uint {
+func (o RollbackBetCancel) Product() int {
 	return o.MessageAttributes.Product
 }
